Permission added for specific acions performed by admin and customer

+++ MoonValleyNursery/src/main/java/com/MVNursery/controller/CustomerController.java
@@ -14,18 +14,26 @@ import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.PutMapping;
 import org.springframework.web.bind.annotation.RequestBody;
+import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.RestController;
 
 import com.MVNursery.exception.CustomerException;
+import com.MVNursery.exception.SessionException;
 import com.MVNursery.model.Customer;
 import com.MVNursery.model.CustomerDTO;
+import com.MVNursery.model.Session;
+import com.MVNursery.model.UserType;
 import com.MVNursery.service.ICustomerService;
+import com.MVNursery.service.ISessionService;
 
 @RestController
 public class CustomerController {
 
 	@Autowired
 	private ICustomerService custService;
+	
+	@Autowired
+	private ISessionService sessionService;
 	
 	@PostMapping("/customers")
 	public ResponseEntity<Customer> addCustomer(@RequestBody @Valid CustomerDTO customer) throws CustomerException{
@@ -34,26 +42,44 @@ public class CustomerController {
 	}
 	
 	@GetMapping("/customers/{customerId}")
-	public ResponseEntity<Customer> getCustomerById(@PathVariable("customerId") Integer customerId) throws CustomerException{
-		Customer customer = custService.getCustomerById(customerId);
-		return new ResponseEntity<>(customer, HttpStatus.ACCEPTED);
+	public ResponseEntity<Customer> getCustomerById(@PathVariable("customerId") Integer customerId, @RequestParam("sessionKey") String sessionKey) throws CustomerException, SessionException{
+		Session session = sessionService.getASessionByKey(sessionKey);
+		if(session.getUserId() == customerId || session.getUserType() == UserType.ADMIN) {
+			Customer customer = custService.getCustomerById(customerId);
+			return new ResponseEntity<>(customer, HttpStatus.ACCEPTED);
+		}
+		throw new SessionException("Please login with the correct credentials");
+		
 	}
 	
 	@DeleteMapping("/customers/{customerId}")
-	public ResponseEntity<Customer> deleteCustomerById(@PathVariable("customerId") Integer customerId) throws CustomerException{
-		Customer customer = custService.deleteCustomerById(customerId);
-		return new ResponseEntity<>(customer, HttpStatus.ACCEPTED);
+	public ResponseEntity<Customer> deleteCustomerById(@PathVariable("customerId") Integer customerId, @RequestParam("sessionKey") String sessionKey) throws CustomerException, SessionException{
+		Session session = sessionService.getASessionByKey(sessionKey);
+		if(session.getUserId() == customerId || session.getUserType() == UserType.ADMIN) {
+			Customer customer = custService.deleteCustomerById(customerId);
+			return new ResponseEntity<>(customer, HttpStatus.ACCEPTED);
+		}
+		throw new SessionException("Please login with the correct credentials");
+		
 	}
 	
 	@PutMapping("/customers")
-	public ResponseEntity<Customer> updateCustomer(@RequestBody @Valid Customer customer) throws CustomerException{
-		Customer updatedCustomer = custService.updateCustomer(customer);
-		return new ResponseEntity<>(updatedCustomer, HttpStatus.ACCEPTED);
+	public ResponseEntity<Customer> updateCustomer(@RequestBody @Valid Customer customer, @RequestParam("SessionKey")String sessionKey) throws CustomerException, SessionException{
+		Session session = sessionService.getASessionByKey(sessionKey);
+		if(session.getUserId() == customer.getCustomerId() || session.getUserType() == UserType.ADMIN) {
+			Customer updatedCustomer = custService.updateCustomer(customer);
+			return new ResponseEntity<>(updatedCustomer, HttpStatus.ACCEPTED);
+		}
+		throw new SessionException("Please login with the correct credentials");
 	}
 	
 	@GetMapping("/customers")
-	public ResponseEntity<List<Customer>> getAllCustomers() throws CustomerException{
-		return new ResponseEntity<>(custService.getAllCustomers(), HttpStatus.ACCEPTED);
+	public ResponseEntity<List<Customer>> getAllCustomers(@RequestParam("sessionKey")String sessionKey) throws CustomerException, SessionException{
+		Session session = sessionService.getASessionByKey(sessionKey);
+		if(session.getUserType() == UserType.ADMIN) {
+			return new ResponseEntity<>(custService.getAllCustomers(), HttpStatus.ACCEPTED);
+		}
+		throw new SessionException("Please login with the correct credentails");
 	}
 }
